requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
openai
pgvector
pdfminer.six
